requirements.txt
asgiref==3.8.1
Django==5.0
django-unfold==0.34.0
sqlparse==0.5.2
typing_extensions==4.12.2
